Display for series is working

+++ src/com/yoharnu/newontv/shows/SeriesDisplay.java
@@ -14,6 +14,7 @@ import android.os.Build;
 import android.os.Bundle;
 import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.app.ProgressDialog;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.ImageView;
@@ -33,25 +34,30 @@ public class SeriesDisplay extends Activity {
 		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
 			getActionBar().setDisplayHomeAsUpEnabled(true);
 		}
-
-		series = new Series(savedInstanceState.getString("series"),
-				Series.ID);
 	}
 
 	protected void onDestroy() {
 		super.onDestroy();
-		for (File f : new File(this.getCacheDir(), "images").listFiles()) {
-			f.delete();
-		}
+		if (new File(this.getCacheDir(), "images").listFiles() != null)
+			for (File f : new File(this.getCacheDir(), "images").listFiles()) {
+				f.delete();
+			}
 		new File(this.getCacheDir(), "images").delete();
 	}
 
 	protected void onStart() {
 		super.onStart();
-		LinearLayout layout = (LinearLayout) findViewById(R.id.layout_series_display);
-		final ImageView image = new ImageView(this);
+
+		final ProgressDialog pd = new ProgressDialog(this);
+		pd.setMessage("Loading...");
+		pd.setIndeterminate(true);
+		pd.setCancelable(false);
+		pd.show();
+		final ImageView image = new ImageView(SeriesDisplay.this);
 		new Thread(new Runnable() {
 			public void run() {
+				series = new Series(SeriesDisplay.this.getIntent().getExtras()
+						.getString("series"), Series.ID);
 				final File imageFile = new File(
 						SeriesDisplay.this.getCacheDir(), "images/"
 								+ series.seriesid);
@@ -70,33 +76,40 @@ public class SeriesDisplay extends Activity {
 				} catch (IOException e) {
 					e.printStackTrace();
 				}
+				SeriesDisplay.this.runOnUiThread(new Runnable() {
+					public void run() {
+						pd.dismiss();
+						LinearLayout layout = (LinearLayout) findViewById(R.id.layout_series_display);
+						layout.addView(image);
+
+						TextView title = new TextView(SeriesDisplay.this);
+						title.setText("Title: " + series.seriesName);
+						layout.addView(title);
+
+						TextView network = new TextView(SeriesDisplay.this);
+						network.setText("Network: " + series.network);
+						layout.addView(network);
+
+						TextView firstAired = new TextView(SeriesDisplay.this);
+						firstAired.setText("First Aired: " + series.firstAired);
+						layout.addView(firstAired);
+
+						TextView status = new TextView(SeriesDisplay.this);
+						status.setText("Status: " + series.status);
+						layout.addView(status);
+
+						TextView type = new TextView(SeriesDisplay.this);
+						type.setText("Type: " + series.classification);
+						layout.addView(type);
+
+						TextView summary = new TextView(SeriesDisplay.this);
+						summary.setText("Summary (courtesy of tvrage.com): "
+								+ series.summary);
+						layout.addView(summary);
+					}
+				});
 			}
 		}).start();
-		layout.addView(image);
-
-		TextView title = new TextView(this);
-		title.setText("Title: " + series.seriesName);
-		layout.addView(title);
-
-		TextView network = new TextView(this);
-		network.setText("Network: " + series.network);
-		layout.addView(network);
-
-		TextView firstAired = new TextView(this);
-		firstAired.setText("First Aired: " + series.firstAired);
-		layout.addView(firstAired);
-
-		TextView status = new TextView(this);
-		status.setText("Status: " + series.status);
-		layout.addView(status);
-
-		TextView type = new TextView(this);
-		type.setText("Type: " + series.classification);
-		layout.addView(type);
-
-		TextView summary = new TextView(this);
-		summary.setText("Summary (courtesy of tvrage.com): " + series.summary);
-		layout.addView(summary);
 	}
 
 	@Override
